requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.6
asn1crypto==0.24.0
astroid==2.1.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
clyent==1.2.2
colorama==0.4.1
conda==4.5.12
cryptography==2.4.2
decorator==4.3.0
Django==2.0.9
django-rest-knox==3.6.0
django-webpack-loader==0.6.0
djangorestframework==3.9.0
entrypoints==0.2.3
Flask==1.0.2
idna==2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.4
mysqlclient==1.3.14
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.4
olefile==0.46
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
psutil==5.4.8
pycosat==0.6.3
pycparser==2.18
Pygments==2.3.1
pylint==2.2.2
pyOpenSSL==18.0.0
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.7
pywin32==223
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
QtPy==1.5.2
requests==2.19.1
ruamel-yaml==0.15.46
Send2Trash==1.5.0
six==1.11.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.23
virtualenv==16.2.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
win-inet-pton==1.0.1
wincertstore==0.2
wrapt==1.10.11
